+++ projects/hipcub/benchmark/benchmark_block_histogram.cpp
@@ -64,13 +64,13 @@ struct histogram
         __shared__ T                                  histogram[BinSize];
         __shared__ typename bhistogram_t::TempStorage storage;
 
-#pragma nounroll
+        _CCCL_PRAGMA_NOUNROLL()
         for(unsigned int trial = 0; trial < Trials; trial++)
         {
             bhistogram_t(storage).Histogram(values, histogram);
         }
 
-#pragma unroll
+        _CCCL_PRAGMA_UNROLL_FULL()
         for(unsigned int offset = 0; offset < BinSize; offset += BlockSize)
         {
             if(offset + hipThreadIdx_x < BinSize)