+++ projects/hipcub/benchmark/benchmark_block_reduce.cpp
@@ -58,7 +58,7 @@ struct reduce
         using breduce_t = hipcub::BlockReduce<T, BlockSize, algorithm>;
         __shared__ typename breduce_t::TempStorage storage;
 
-#pragma nounroll
+        _CCCL_PRAGMA_NOUNROLL()
         for(unsigned int trial = 0; trial < Trials; trial++)
         {
             reduced_value = breduce_t(storage).Reduce(values, hipcub::Sum());