+++ projects/hipcub/benchmark/benchmark_block_scan.cpp
@@ -57,7 +57,7 @@ struct inclusive_scan
         using bscan_t = hipcub::BlockScan<T, BlockSize, Algorithm>;
         __shared__ typename bscan_t::TempStorage storage;
 
-#pragma nounroll
+        _CCCL_PRAGMA_NOUNROLL()
         for(unsigned int trial = 0; trial < Trials; trial++)
         {
             bscan_t(storage).InclusiveScan(values, values, hipcub::Sum());
@@ -87,7 +87,7 @@ struct exclusive_scan
         using bscan_t = hipcub::BlockScan<T, BlockSize, Algorithm>;
         __shared__ typename bscan_t::TempStorage storage;
 
-#pragma nounroll
+        _CCCL_PRAGMA_NOUNROLL()
         for(unsigned int trial = 0; trial < Trials; trial++)
         {
             bscan_t(storage).ExclusiveScan(values, values, init, hipcub::Sum());