+++ projects/hipcub/benchmark/benchmark_warp_load.cpp
@@ -50,7 +50,7 @@ __device__ auto warp_load_benchmark(T* d_input, T* d_output)
 
     WarpLoadT(temp_storage[warp_id]).Load(d_input + global_warp_id * tile_size, thread_data);
 
-#pragma unroll
+    _CCCL_PRAGMA_UNROLL_FULL()
     for(unsigned i = 0; i < ItemsPerThread; ++i)
     {
         const unsigned striped_global_idx