+++ projects/hipcub/hipcub/include/hipcub/backend/rocprim/block/block_merge_sort.hpp
@@ -95,7 +95,7 @@ HIPCUB_DEVICE __forceinline__ void SerialMerge(KeyT *keys_shared,
    KeyT key1 = keys_shared[keys1_beg];
    KeyT key2 = keys_shared[keys2_beg];
 
-#pragma unroll
+   _CCCL_PRAGMA_UNROLL_FULL()
    for (int item = 0; item < ITEMS_PER_THREAD; ++item)
    {
        bool p = (keys2_beg < keys2_end) &&
@@ -387,7 +387,7 @@ public:
       //
       KeyT max_key = oob_default;
 
-      #pragma unroll
+      _CCCL_PRAGMA_UNROLL_FULL()
       for (int item = WARP_SORT ? 1 : 0; item < ITEMS_PER_THREAD; ++item)
       {
         if (ITEMS_PER_THREAD * static_cast<int>(linear_tid) + item < valid_items)
@@ -411,7 +411,7 @@ public:
     // each thread has sorted keys
     // merge sort keys in shared memory
     //
-    #pragma unroll
+    _CCCL_PRAGMA_UNROLL_FULL()
     for (int target_merged_threads_number = 2;
          target_merged_threads_number <= NUM_THREADS;
          target_merged_threads_number *= 2)
@@ -423,7 +423,7 @@ public:
 
       // store keys in shmem
       //
-      #pragma unroll
+      _CCCL_PRAGMA_UNROLL_FULL()
       for (int item = 0; item < ITEMS_PER_THREAD; ++item)
       {
         int idx                       = ITEMS_PER_THREAD * linear_tid + item;
@@ -482,7 +482,7 @@ public:
 
         // store keys in shmem
         //
-        #pragma unroll
+        _CCCL_PRAGMA_UNROLL_FULL()
         for (int item = 0; item < ITEMS_PER_THREAD; ++item)
         {
           int idx = ITEMS_PER_THREAD * linear_tid + item;
@@ -493,7 +493,7 @@ public:
 
         // gather items from shmem
         //
-        #pragma unroll
+        _CCCL_PRAGMA_UNROLL_FULL()
         for (int item = 0; item < ITEMS_PER_THREAD; ++item)
         {
           items[item] = temp_storage.items_shared[indices[item]];