+++ projects/hipcub/test/hipcub/test_hipcub_single_pass_scan_operators.cpp
@@ -272,7 +272,7 @@ static void RunningPrefixKernel(T* d_input, T* d_output)
 
     prefix_type prefix(T(), ScanOp{});
 
-#pragma unroll
+    _CCCL_PRAGMA_UNROLL_FULL()
     for(int i = 0; i < num_items; ++i)
     {
         T value            = d_input[i];